# perf: add caching for extension getter to enhance performance