--- a/application/src/main/java/run/halo/app/plugin/extensionpoint/DefaultExtensionGetter.java
+++ b/application/src/main/java/run/halo/app/plugin/extensionpoint/DefaultExtensionGetter.java
@@ -1,7 +1,5 @@
 package run.halo.app.plugin.extensionpoint;
 
-import static run.halo.app.extension.index.query.QueryFactory.equal;
-
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
@@ -11,15 +9,9 @@ import org.pf4j.ExtensionPoint;
 import org.pf4j.PluginManager;
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.core.annotation.AnnotationAwareOrderComparator;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import run.halo.app.extension.ListOptions;
-import run.halo.app.extension.ListResult;
-import run.halo.app.extension.PageRequestImpl;
-import run.halo.app.extension.ReactiveExtensionClient;
-import run.halo.app.extension.router.selector.FieldSelector;
 import run.halo.app.infra.SystemConfigurableEnvironmentFetcher;
 import run.halo.app.infra.SystemSetting.ExtensionPointEnabled;
 
@@ -33,7 +25,9 @@ public class DefaultExtensionGetter implements ExtensionGetter {
 
     private final BeanFactory beanFactory;
 
-    private final ReactiveExtensionClient client;
+    private final ExtensionDefinitionGetter extensionDefinitionGetter;
+
+    private final ExtensionPointDefinitionGetter extensionPointDefinitionGetter;
 
     @Override
     public <T extends ExtensionPoint> Flux<T> getExtensions(Class<T> extensionPoint) {
@@ -86,9 +80,7 @@ public class DefaultExtensionGetter implements ExtensionGetter {
                 }
                 var extensions = getExtensions(extensionPoint).cache();
                 return Flux.fromIterable(extensionDefNames)
-                    .flatMapSequential(extensionDefName ->
-                        client.fetch(ExtensionDefinition.class, extensionDefName)
-                    )
+                    .flatMapSequential(extensionDefinitionGetter::get)
                     .flatMapSequential(extensionDef -> {
                         var className = extensionDef.getSpec().getClassName();
                         return extensions.filter(
@@ -101,15 +93,7 @@ public class DefaultExtensionGetter implements ExtensionGetter {
 
     private Mono<ExtensionPointDefinition> fetchExtensionPointDefinition(
         Class<? extends ExtensionPoint> extensionPoint) {
-        var listOptions = new ListOptions();
-        listOptions.setFieldSelector(FieldSelector.of(
-            equal("spec.className", extensionPoint.getName())
-        ));
-        var sort = Sort.by("metadata.creationTimestamp", "metadata.name").ascending();
-        return client.listBy(ExtensionPointDefinition.class, listOptions,
-                PageRequestImpl.ofSize(1).withSort(sort)
-            )
-            .flatMap(list -> Mono.justOrEmpty(ListResult.first(list)));
+        return extensionPointDefinitionGetter.getByClassName(extensionPoint.getName());
     }
 
 }