--- a/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionDefinitionGetter.java
+++ b/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionDefinitionGetter.java
@@ -1,0 +1,13 @@
+package run.halo.app.plugin.extensionpoint;
+
+import reactor.core.publisher.Mono;
+
+public interface ExtensionDefinitionGetter {
+
+    /**
+     * Gets extension definition by extension definition name.
+     *
+     * @param name extension definition name
+     */
+    Mono<ExtensionDefinition> get(String name);
+}