--- a/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionPointDefinitionGetter.java
+++ b/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionPointDefinitionGetter.java
@@ -1,0 +1,14 @@
+package run.halo.app.plugin.extensionpoint;
+
+import reactor.core.publisher.Mono;
+
+public interface ExtensionPointDefinitionGetter {
+
+    /**
+     * Gets extension point definition by extension point class.
+     * <p>Retrieve by filedSelector: <code>spec.className</code></p>
+     *
+     * @param className extension point class name
+     */
+    Mono<ExtensionPointDefinition> getByClassName(String className);
+}