--- a/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionPointDefinitionGetterImpl.java
+++ b/application/src/main/java/run/halo/app/plugin/extensionpoint/ExtensionPointDefinitionGetterImpl.java
@@ -1,0 +1,26 @@
+package run.halo.app.plugin.extensionpoint;
+
+import org.springframework.stereotype.Component;
+import reactor.core.publisher.Mono;
+import run.halo.app.extension.ExtensionClient;
+
+@Component
+public class ExtensionPointDefinitionGetterImpl
+    extends AbstractDefinitionGetter<ExtensionPointDefinition>
+    implements ExtensionPointDefinitionGetter {
+
+    public ExtensionPointDefinitionGetterImpl(ExtensionClient client) {
+        super(client, new ExtensionPointDefinition());
+    }
+
+    @Override
+    public Mono<ExtensionPointDefinition> getByClassName(String className) {
+        return Mono.fromSupplier(() -> cache.get(className));
+    }
+
+    @Override
+    void putCache(ExtensionPointDefinition definition) {
+        var className = definition.getSpec().getClassName();
+        cache.put(className, definition);
+    }
+}